Extract engine implementation module
